requirements.txt
annotated-types==0.7.0
anyio==4.11.0
brotlicffi @ file:///opt/miniconda3/conda-bld/brotlicffi_1761745169121/work
cachetools==6.2.1
certifi==2025.10.5
cffi @ file:///opt/miniconda3/conda-bld/cffi_1760098029095/work
charset-normalizer @ file:///opt/miniconda3/conda-bld/charset-normalizer_1761744843696/work
contourpy==1.3.0
cycler==0.12.1
exceptiongroup==1.3.0
filelock @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_f6dmuewahq/croot/filelock_1744281400993/work
fonttools==4.60.1
fsspec==2025.10.0
gmpy2 @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_51391juwln/croot/gmpy2_1738085477864/work
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
importlib_resources==6.5.2
Jinja2 @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_2cnn4kenrm/croot/jinja2_1741710859444/work
kiwisolver==1.4.7
MarkupSafe @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_1f_uj4vxik/croot/markupsafe_1738584045311/work
matplotlib==3.9.4
mpmath @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_17iu6a8a3m/croot/mpmath_1690848269369/work
networkx @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_797xfkx_y7/croot/networkx_1717597499877/work
numpy @ file:///Users/ktietz/demo/mc3/conda-bld/numpy_1629114958789/work
ollama==0.6.0
opencv-python==4.12.0.88
packaging==25.0
pillow @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_bfoqpm49mx/croot/pillow_1744613079607/work
polars==1.35.1
polars-runtime-32==1.35.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybullet @ file:///Users/runner/miniforge3/conda-bld/bullet_1747514855015/work
pycparser @ file:///opt/miniconda3/conda-bld/pycparser_1757496095003/work
pydantic==2.12.4
pydantic_core==2.41.5
pyparsing==3.2.5
PySocks @ file:///opt/miniconda3/conda-bld/pysocks_1761753009935/work
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_faoex52hrr/croot/pyyaml_1728657970485/work
requests @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_ef59rwwti1/croot/requests_1756709671417/work
rsa==4.9.1
scipy==1.13.1
six==1.17.0
sniffio==1.3.1
sympy @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_9b15dnv4zk/croot/sympy_1756713487054/work
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.3.225
ultralytics-thop==2.0.18
uritemplate==4.2.0
urllib3 @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_95v32ztwo_/croot/urllib3_1750775480393/work
zipp==3.23.0
